requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-elasticsearch-dsl==0.5.1
django-extensions==2.1.9
-e git+https://github.com/django-haystack/django-haystack.git@802b0f6f4b3b99314453261876a32bac2bbec94f#egg=django_haystack
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
djangorestframework==3.9.4
elasticsearch==5.5.3
elasticsearch-dsl==5.4.0
idna==2.8
ipaddress==1.0.22
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.1.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.2
